refactor: fix component splitting of comments
